@@ -6,7 +6,7 @@ import 'Widgets/AppBarCustom.dart';
 final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
 
 class CandidatScreen extends StatelessWidget {
-  CandidatScreen({Key key}) : super(key: key);
+  CandidatScreen({Key? key}) : super(key: key);
 
   final scaffoldKey = GlobalKey<ScaffoldState>();
 
@@ -14,7 +14,7 @@ class CandidatScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       key: scaffoldKey,
-      appBar: AppBarCustom(scaffoldKey:scaffoldKey),
+      appBar: AppBarCustom(scaffoldKey:scaffoldKey, key: null,),
       backgroundColor: Colors.white,
       body:  SafeArea(
         child: SingleChildScrollView(
