@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 class AppBarCustom extends StatelessWidget implements PreferredSizeWidget {
   final GlobalKey<ScaffoldState> scaffoldKey;
 
-  const AppBarCustom({Key key, @required this.scaffoldKey}) : super(key: key);
+  const AppBarCustom({required Key? key, required this.scaffoldKey}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -13,7 +13,7 @@ class AppBarCustom extends StatelessWidget implements PreferredSizeWidget {
       leading: IconButton(
         icon: Icon(Icons.menu),
         onPressed: () {
-          if (scaffoldKey.currentState?.isDrawerOpen) {
+          if (scaffoldKey.currentState!.isDrawerOpen) {
             scaffoldKey.currentState?.openEndDrawer();
           } else {
             scaffoldKey.currentState?.openDrawer();
